% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaar-pkg.R
\docType{package}
\name{aaaaR-pkg}
\alias{aaaaR-pkg}
\alias{aaaaR-pkg-package}
\title{Let the aminoacids do the talking, in R}
\description{
aaaaR provides a rapid interface to translate your words into
aminoacids. And as a side effect, it leaves you wondering why noone did it
until now.
}
\author{
Federico Marini \email{marinif@uni-mainz.de}

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}
