% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaar.R
\name{aaaar}
\alias{aaaar}
\title{aaaar}
\usage{
aaaar(mystring = "damn")
}
\arguments{
\item{mystring}{A character string}
}
\value{
A character string where the letters composing your string have
been replaced by aminoacids - where possible
}
\description{
Transform your input string character into a string composed by aminoacids
}
\examples{
aaaar("damn")
}
