% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaar-pkg.R
\docType{data}
\name{aatable}
\alias{aatable}
\title{A table containing the correspondences between letters and aminoacids}
\description{
Source: Wikipedia entry
}
\author{
Federico Marini \email{marinif@uni-mainz.de}
}
